+++ src/environment_variables/env_free.c
@@ -6,7 +6,27 @@
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 02:28:59 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/09 02:28:59 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/09 02:45:54 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "env.h"
+#include <stdlib.h>
+
+void	env_free(t_env **stack)
+{
+	t_env	*node;
+	t_env	*next_node;
+
+	if (!stack || !(*stack))
+		return ;
+	node = *stack;
+	while (node)
+	{
+		next_node = node->next;
+		free(node->name_value);
+		free(node);
+		node = next_node;
+	}
+	*stack = NULL;
+}
