+++ src/environment_variables/env_get_value_by_name.c
@@ -6,7 +6,28 @@
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 02:27:08 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/09 02:27:09 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/09 02:50:27 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "env.h"
+#include "libft.h"
+#include <stddef.h>
+
+char	*env_get_value_by_name(t_env *stack, const char *name)
+{
+	t_env	*node;
+	size_t	i;
+
+	node = stack;
+	while (node)
+	{
+		i = 0;
+		while (node->name_value[i] == name[i])
+			i++;
+		if (node->name_value[i] == '=' && name[i] == '\0')
+			return (ft_strdup(&(node->name_value[i + 1])));
+		node = node->next;
+	}
+	return (NULL);
+}
